sprite_tools/requirements.txt
pygame==2.5.1
pillow==9.3.0